+++ freeway-http/src/main/java/com/jujin/freeway/http/engine/ws/WebSocketUtils.java
@@ -5,9 +5,9 @@ import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.util.Base64;
 
-public final class WebSocketUtil {
+public final class WebSocketUtils {
 
-    private WebSocketUtil() {}
+    private WebSocketUtils() {}
 
     private static final String MAGIC = "258EAFA5-E914-47DA-95CA-C5AB0DC85B11";